fix updateAxisScaleLabel
